+++ packages/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.2.1] - 2026-07-01
+
+### Added
+
+- Preserve deleted local items.
+- Generate `FileMetadata` types for file fields.
+
 ## [0.2.0] - 2026-06-30
 
 ### Changed